+++ src/dispatcher/config.py
@@ -9,6 +9,8 @@ class DispatcherSettings(BaseSettings):
     nats_user: str
     # Nats password specific to Dispatcher
     nats_password: str
+    # Maximum number of reconnect attempts to the NATS server
+    max_reconnect_attempts: int = 5
 
     # Dispatcher: username
     dispatcher_username: str