+++ tests/e2e/test_consumer_client.py
@@ -38,9 +38,9 @@ async def test_get_empty_messages(
 async def test_send_message(
     provisioning_client: shared.client.AsyncClient,
     simple_subscription: str,
-    provisioning_base_url: str,
+    provisioning_api_base_url: str,
 ):
-    data = create_message_via_events_api(provisioning_base_url)
+    data = create_message_via_events_api(provisioning_api_base_url)
 
     response = await provisioning_client.get_subscription_messages(
         name=simple_subscription,